scanner/requirements.txt
# FastAPI
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# Task Queue
celery>=5.3.0
redis>=5.0.0

# HTTP Client
httpx>=0.25.0
requests>=2.31.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Data Processing
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dateutil>=2.8.0

# Report Generation
jinja2>=3.1.0
weasyprint>=59.0
markupsafe>=2.1.0

# Utilities
python-magic>=0.4.27
filetype>=1.2.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0
